# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.AuthResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/domain.User'
    type: object
  domain.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  domain.ForgotPasswordRequest:
    properties:
      email:
        example: john@lissanai.com
        type: string
    required:
    - email
    type: object
...
  domain.LoginRequest:
    properties:
      email:
        example: john@lissanai.com
        type: string
      password:
        example: strongpassword123
        type: string
    required:
    - email
    - password
    type: object
  domain.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  domain.PushTokenRequest:
    properties:
      platform:
        example: ios
        type: string
      token:
        example: fcm_token_123
        type: string
    required:
    - platform
    - token
    type: object
...
  domain.RefreshTokenRequest:
    properties:
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    required:
    - refresh_token
    type: object
  domain.RegisterRequest:
    properties:
      email:
        example: john@lissanai.com
        type: string
      name:
        example: John Doe
        type: string
      password:
        example: strongpassword123
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
  domain.ResetPasswordRequest:
    properties:
      new_password:
        example: newstrongpassword123
        minLength: 8
        type: string
      token:
        example: reset_token_123
        type: string
    required:
    - new_password
    - token
    type: object
  domain.SocialAuthRequest:
    properties:
      access_token:
        example: ya29.a0AfH6SMC...
        type: string
      email:
        example: john@lissanai.com
        type: string
      name:
        example: John Doe
        type: string
      provider:
        example: google
        type: string
    required:
    - access_token
    - provider
    type: object
...
  domain.TokenResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
    type: object
  domain.UpdateProfileRequest:
    properties:
      name:
        example: John Updated
        type: string
      settings:
        additionalProperties: true
        type: object
    type: object
  domain.User:
    properties:
      created_at:
        type: string
      current_streak:
        description: Streak System
        type: integer
      email:
        type: string
      freeze_count:
        type: integer
      id:
        type: string
      last_activity_date:
        type: string
      longest_streak:
        type: integer
      name:
        type: string
      provider:
        type: string
      settings:
        additionalProperties: true
        type: object
      streak_frozen:
        type: boolean
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: AI-powered English coach for Ethiopians seeking global job opportunities
  title: LissanAI API
  version: "1.0"
paths:
...
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Send password reset link to user's email
      parameters:
      - description: User Email
        in: body
        name: email
        required: true
        schema:
          $ref: '#/definitions/domain.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Request password reset
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password
      parameters:
      - description: User Login Credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/domain.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Login user
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Invalidate user's session token
      parameters:
      - description: Refresh Token (optional)
        in: body
        name: refreshToken
        schema:
          $ref: '#/definitions/domain.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Use refresh token to get a new access token
      parameters:
      - description: Refresh Token
        in: body
        name: refreshToken
        required: true
        schema:
          $ref: '#/definitions/domain.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Refresh access token
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account with email and password
      parameters:
      - description: User Registration Information
        in: body
        name: userInfo
        required: true
        schema:
          $ref: '#/definitions/domain.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Register a new user
      tags:
      - Auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Set new password using reset token
      parameters:
      - description: Password Reset Information
        in: body
        name: resetInfo
        required: true
        schema:
          $ref: '#/definitions/domain.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Reset password
      tags:
      - Auth
  /auth/social:
    post:
      consumes:
      - application/json
      description: Authenticate or register user using social providers (Google, Apple)
      parameters:
      - description: Social Authentication Information
        in: body
        name: socialInfo
        required: true
        schema:
          $ref: '#/definitions/domain.SocialAuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Social authentication
      tags:
      - Auth
...
  /users/me:
    delete:
      description: Allow a user to delete their account
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user account
      tags:
      - Users
    get:
      description: Get the profile of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Update parts of the user's profile (name, settings)
      parameters:
      - description: Profile Update Information
        in: body
        name: profileInfo
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - Users
  /users/me/push-token:
    post:
      consumes:
      - application/json
      description: Register a device token (FCM/APNs) for push notifications
      parameters:
      - description: Push Token Information
        in: body
        name: tokenInfo
        required: true
        schema:
          $ref: '#/definitions/domain.PushTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Register push token
      tags:
      - Users
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"